# public/swagger.json
{
	"components": {
...
		"schemas": {
...
			"CourtSurface": {
				"enum": [
					"GRASS",
					"CLAY",
					"HARD"
				],
				"type": "string"
			},
			"ClubResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"address": {
						"type": "string"
					},
					"description": {
						"type": "string"
					},
					"city": {
						"type": "string"
					},
					"country": {
						"type": "string"
					},
					"courts": {
						"items": {
							"$ref": "#/components/schemas/CourtResponse"
						},
						"type": "array"
					},
					"courtsNumber": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"address",
					"description",
					"city",
					"country",
					"courtsNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CourtResponse": {
				"properties": {
					"name": {
						"type": "string"
					},
					"surface": {
						"$ref": "#/components/schemas/CourtSurface"
					},
					"stands": {
						"type": "boolean"
					},
					"roof": {
						"type": "boolean"
					},
					"double": {
						"type": "boolean"
					},
					"clubId": {
						"type": "string"
					},
					"club": {
						"$ref": "#/components/schemas/ClubResponse"
					},
					"pricePerHour": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"surface",
					"stands",
					"roof",
					"double",
					"clubId",
					"pricePerHour"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CourtCreateRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"surface": {
						"$ref": "#/components/schemas/CourtSurface"
					},
					"stands": {
						"type": "boolean"
					},
					"roof": {
						"type": "boolean"
					},
					"double": {
						"type": "boolean"
					},
					"pricePerHour": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"surface",
					"stands",
					"roof",
					"double",
					"pricePerHour"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AssignCourtRequest": {
				"properties": {
					"courtEntityID": {
						"type": "string"
					},
					"clubEntityID": {
						"type": "string"
					}
				},
				"required": [
					"courtEntityID",
					"clubEntityID"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateCourtRequest": {
				"properties": {
					"name": {
						"type": "string"
					},
					"surface": {
						"$ref": "#/components/schemas/CourtSurface"
					},
					"stands": {
						"type": "boolean"
					},
					"roof": {
						"type": "boolean"
					},
					"double": {
						"type": "boolean"
					},
					"pricePerHour": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"name",
					"surface",
					"stands",
					"roof",
					"double",
					"pricePerHour"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"CreateEnemyRequest": {
				"properties": {
					"bookingEntityID": {
						"type": "string"
					},
					"playerEntityID": {
						"type": "string"
					},
					"numberOfPlayersNeeded": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"bookingEntityID",
					"playerEntityID",
					"numberOfPlayersNeeded"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AcceptEnemyRequest": {
				"properties": {
					"requestEntityID": {
						"type": "string"
					},
					"playerEntityID": {
						"type": "string"
					}
				},
				"required": [
					"requestEntityID",
					"playerEntityID"
				],
				"type": "object",
				"additionalProperties": false
			},
			"EnemyRequestResponse": {
				"properties": {
					"entityId": {
						"type": "string"
					},
					"bookingEntityID": {
						"type": "string"
					},
					"courtEntityId": {
						"type": "string"
					},
					"date": {
						"type": "string",
						"format": "date-time"
					},
					"totalPrice": {
						"type": "number",
						"format": "double"
					},
					"from": {
						"type": "number",
						"format": "double"
					},
					"to": {
						"type": "number",
						"format": "double"
					},
					"playerNickname": {
						"type": "string"
					},
					"numberOfPlayersNeeded": {
						"type": "number",
						"format": "double"
					},
					"acceptedPlayerNicnames": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"active": {
						"type": "boolean"
					},
					"courtName": {
						"type": "string"
					},
					"courtSurface": {
						"type": "string"
					}
				},
				"required": [
					"entityId",
					"bookingEntityID",
					"courtEntityId",
					"date",
					"totalPrice",
					"from",
					"to",
					"playerNickname",
					"numberOfPlayersNeeded",
					"acceptedPlayerNicnames",
					"active",
					"courtName",
					"courtSurface"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateEnemyRequest": {
				"properties": {
					"bookingEntityID": {
						"type": "string"
					},
					"numberOfPlayersNeeded": {
						"type": "number",
						"format": "double"
					},
					"acceptedBy": {
						"items": {
							"type": "string"
						},
						"type": "array"
					}
				},
				"required": [
					"bookingEntityID",
					"numberOfPlayersNeeded",
					"acceptedBy"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"jwt": {
				"type": "apiKey",
				"name": "Authorization",
				"in": "header"
			}
		}
	},
	"info": {
		"title": "tennis",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"openapi": "3.0.0",
	"paths": {
...
		"/courts/all": {
			"get": {
				"operationId": "GetAllCourts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CourtResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Courts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/courts/{entityId}": {
			"get": {
				"operationId": "GetById",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/CourtResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Courts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteCourt",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Courts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateCourt",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Courts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateCourtRequest"
							}
						}
					}
				}
			}
		},
		"/courts": {
			"post": {
				"operationId": "CreateCourt",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {}
							}
						}
					}
				},
				"tags": [
					"Courts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CourtCreateRequest"
							}
						}
					}
				}
			}
		},
		"/courts/unassigned": {
			"get": {
				"operationId": "GetUnassignedCourts",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CourtResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Courts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/courts/assign": {
			"post": {
				"operationId": "AssignCourtToClub",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"tags": [
					"Courts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AssignCourtRequest"
							}
						}
					}
				}
			}
		},
		"/courts/price": {
			"get": {
				"operationId": "FindByPrice",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/CourtResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Courts"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "query",
						"name": "from",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "to",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/requests": {
			"post": {
				"operationId": "CreateRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateEnemyRequest"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetActiveRequests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/EnemyRequestResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/requests/accept": {
			"post": {
				"operationId": "AcceptRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/AcceptEnemyRequest"
							}
						}
					}
				}
			}
		},
		"/requests/inactive": {
			"get": {
				"operationId": "GetInactiveRequests",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/EnemyRequestResponse"
									},
									"type": "array"
								}
							}
						}
					}
				},
				"tags": [
					"Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": []
			}
		},
		"/requests/{entityId}": {
			"delete": {
				"operationId": "DeleteEnemyRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"get": {
				"operationId": "GetEnemyRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/EnemyRequestResponse"
								}
							}
						}
					}
				},
				"tags": [
					"Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"patch": {
				"operationId": "UpdateEnemyRequest",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"type": "string"
								}
							}
						}
					}
				},
				"tags": [
					"Requests"
				],
				"security": [
					{
						"jwt": []
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "entityId",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateEnemyRequest"
							}
						}
					}
				}
			}
		}
	},
	"servers": [
		{
			"url": "/"
		}
	]
}